# docs/swagger.yaml
definitions:
...
  shortpress-server_internal_api.IDListResponseData-string:
    properties:
      hasMore:
        description: Whether there are more items
        example: true
        type: boolean
      items:
        description: ID information list
        items:
          type: string
        type: array
      page:
        description: Current page number
        example: 1
        type: integer
      pageSize:
        description: Items per page
        example: 20
        type: integer
      total:
        description: Total number of videos
        example: 100
        type: integer
    type: object
...
  shortpress-server_internal_api.InternalGetVideoConfigResponse:
    properties:
      config:
        description: Config JSON data
        items:
          type: integer
        type: array
      vid:
        description: Video ID
        example: video-123
        type: string
    type: object
...
  shortpress-server_internal_api.Response:
    properties:
      code:
        type: integer
      data: {}
      info:
        type: string
    type: object
...
  shortpress-server_internal_api.VideoAddSourcesRequest:
    properties:
      sources:
        description: Sources to add
        items:
          $ref: '#/definitions/shortpress-server_internal_api.VideoSourceCreateItem'
        type: array
      vid:
        description: Target video id
        type: string
    required:
    - sources
    - vid
    type: object
  shortpress-server_internal_api.VideoAddSourcesResponse:
    properties:
      code:
        type: integer
      data:
        properties:
          vid:
            type: string
        type: object
      info:
        type: string
    type: object
  shortpress-server_internal_api.VideoDeleteRequest:
    properties:
      vids:
        description: List of video IDs to delete
        example:
        - '["vid-123"'
        - '"vid-456"]'
        items:
          type: string
        type: array
    required:
    - vids
    type: object
...
  shortpress-server_internal_api.VideoInfo:
    properties:
      config:
        description: Custom configuration (arbitrary JSON)
        items:
          type: integer
        type: array
      cover:
        description: Video cover
        example: http://example.com/cover.jpg
        type: string
      createdAt:
        description: Creation time
        example: 1609459200
        type: integer
      description:
        description: Video description
        example: This is an example video
        type: string
      duration:
        description: Video duration (seconds)
        example: 180
        type: integer
      seo:
        allOf:
        - $ref: '#/definitions/shortpress-server_internal_api.VideoSeo'
        description: Video SEO information
      sources:
        description: 'New: multiple playback sources'
        items:
          $ref: '#/definitions/shortpress-server_internal_api.VideoSourceInfo'
        type: array
      status:
        description: 'Video status: 1-Unpublished, 2-Published, 3-Offline, 127-Deleted'
        example: 1
        type: integer
      subtitles:
        allOf:
        - $ref: '#/definitions/shortpress-server_internal_types.SubtitleTracks'
        description: Optional subtitle tracks specific to this source
      tags:
        description: Video tags
        example: example,tutorial
        type: string
      title:
        description: Video title
        example: Example video
        type: string
      updatedAt:
        description: Update time
        example: 1609459200
        type: integer
      vid:
        description: Unique video identifier
        example: vid-123456
        type: string
    type: object
...
  shortpress-server_internal_api.VideoListResponseData:
    properties:
      items:
        description: Video list
        items:
          $ref: '#/definitions/shortpress-server_internal_api.VideoInfo'
        type: array
      page:
        description: Current page number
        example: 1
        type: integer
      pageSize:
        description: Items per page
        example: 20
        type: integer
      total:
        description: Total count
        example: 100
        type: integer
    type: object
  shortpress-server_internal_api.VideoReplaceData:
    properties:
      cover:
        description: Video cover
        type: string
      vid:
        description: Unique video identifier
        type: string
      videoSourceUrl:
        description: Video source play URL
        type: string
    type: object
  shortpress-server_internal_api.VideoReplaceResonse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/shortpress-server_internal_api.VideoReplaceData'
      info:
        type: string
    type: object
  shortpress-server_internal_api.VideoSeo:
    properties:
      Keywords:
        description: Keywords
        example: example,tutorial
        type: string
      description:
        description: Video description
        example: This is an example video
        type: string
      title:
        description: Video title
        example: Example video
        type: string
    type: object
...
  shortpress-server_internal_api.VideoUploadResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/shortpress-server_internal_api.VideoUploadResponseData'
      info:
        type: string
    type: object
  shortpress-server_internal_api.VideoUploadResponseData:
    properties:
      vids:
        description: List of successfully uploaded video IDs
        example:
        - '["vid-123"'
        - '"vid-456"]'
        items:
          type: string
        type: array
    type: object
...
  shortpress-server_internal_types.SubtitleTracks:
    additionalProperties:
      $ref: '#/definitions/shortpress-server_internal_types.SubtitleItem'
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: shortpress
    url: http://www.swagger.io/support
  description: This is shortpress server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Shortpress Server API
  version: "0.8"
paths:
...
  /api/internal/video/config:
    get:
      consumes:
      - application/json
      description: Internal service-to-service API for getting video configuration.
        No authentication required.
      parameters:
      - description: Video ID
        in: query
        name: vid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Config retrieved successfully
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.InternalGetVideoConfigResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
        "404":
          description: Video not found
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      summary: 'Internal API: Get video config'
      tags:
      - internal-video
...
  /api/video/batch-get:
    get:
      consumes:
      - application/json
      description: Query brief video information by batch video IDs (note quantity
        control...)
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        type: string
      - description: Video ID list, comma separated
        example: '"vid-123456,vid-123457"'
        in: query
        name: vids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns brief video information list
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.VideoListResponseData'
      security:
      - Bearer: []
      summary: Batch get brief video information
      tags:
      - video
  /api/video/delete:
    post:
      consumes:
      - application/json
      description: Batch delete specified videos
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Video deletion parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shortpress-server_internal_api.VideoDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Deletion successful
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      security:
      - Bearer: []
      summary: Delete videos
      tags:
      - video
  /api/video/get:
    get:
      consumes:
      - application/json
      description: Query video details by video ID
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Video ID
        example: '"vid-123456"'
        in: query
        name: vid
        required: true
        type: string
      - default: false
        description: Whether SEO information is needed
        in: query
        name: need_seo
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Returns detailed video information
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.VideoInfo'
      security:
      - Bearer: []
      summary: Get video details
      tags:
      - video
  /api/video/list:
    get:
      consumes:
      - application/json
      description: Search videos by keyword, supports pagination
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Sort type (0:created time desc 1:name sort)
        example: 1
        in: query
        name: sortType
        type: integer
      - description: Video status
        example: 1
        in: query
        name: status
        type: integer
      - description: Video upload status, comma separated, supports multiple status
          combinations
        in: query
        name: uploadStatus
        type: integer
      - description: Page number, default 1
        example: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Page size, default 10
        example: 20
        in: query
        maximum: 20
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns video list in playlist
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.IDListResponseData-string'
      security:
      - Bearer: []
      summary: Display all videos of current creator
      tags:
      - video
  /api/video/modify:
    post:
      consumes:
      - application/json
      description: Modify basic information of a specified video
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Video modification parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shortpress-server_internal_api.VideoInfo'
      produces:
      - application/json
      responses:
        "200":
          description: Modification successful
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      security:
      - Bearer: []
      summary: Modify video information
      tags:
      - video
  /api/video/replace:
    post:
      consumes:
      - multipart/form-data
      description: Synchronously replace a video
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Video ID to replace
        example: '"vid-123456"'
        in: query
        name: vid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Replacement successful
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.VideoReplaceResonse'
      security:
      - Bearer: []
      summary: Replace video, keeping the same video ID
      tags:
      - video
  /api/video/search:
    get:
      consumes:
      - application/json
      description: Search videos by keyword, supports pagination
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Playlist ID, if not provided, search all videos
        example: playlistid
        in: query
        name: playlistId
        type: string
      - description: Exclude playlist id, if not provided, search all videos
        example: playlistid
        in: query
        name: excludePlaylistId
        type: string
      - description: Sort type (0:created time desc 1:name sort)
        example: 1
        in: query
        name: orderType
        type: integer
      - description: Video status
        example: 1
        in: query
        name: status
        type: integer
      - description: Video upload status, comma separated, supports multiple status
          combinations
        in: query
        name: uploadStatus
        type: integer
      - description: Page number, default 1
        example: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Page size, default 10
        example: 20
        in: query
        maximum: 20
        minimum: 1
        name: pageSize
        type: integer
      - description: Search keyword
        example: tutorial
        in: query
        name: keyword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns search results
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.IDListResponseData-string'
      security:
      - Bearer: []
      summary: Search videos
      tags:
      - video
  /api/video/sources/add:
    post:
      consumes:
      - application/json
      description: Add one or multiple non-local sources for a video
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add sources request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shortpress-server_internal_api.VideoAddSourcesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Added successfully
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.VideoAddSourcesResponse'
      security:
      - Bearer: []
      summary: Add network sources for a video
      tags:
      - video
  /api/video/upload:
    post:
      consumes:
      - multipart/form-data
      description: Batch upload video files, returns a list of created video VIDs
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Playlist ID, if not provided, videos won't be added to any playlist
        example: playlistid
        in: formData
        name: playlistId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns a list of uploaded video VIDs
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.VideoUploadResponse'
      security:
      - Bearer: []
      summary: Batch upload videos
      tags:
      - video
  /api/video/upload-subtitle:
    post:
      consumes:
      - multipart/form-data
      description: Upload video subtitle file, supports SRT and VTT formats
      parameters:
      - description: Bearer user token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Video ID
        example: '"vid-123456"'
        in: formData
        name: vid
        required: true
        type: string
      - description: Subtitle file, supports SRT and VTT formats
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Returns the subtitle URL
          schema:
            $ref: '#/definitions/shortpress-server_internal_api.Response'
      security:
      - Bearer: []
      summary: Upload video subtitle
      tags:
      - video
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"